pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Felipe Kutz <felipekutz@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-taggit = "^4.0.0"
markdown = "^3.4.4"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
model-bakery = "^1.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
